+++ security-service/src/main/java/com/gitlab/SecurityMain.java
@@ -1,7 +1,14 @@
 package com.gitlab;
 
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+
+@SpringBootApplication
+@EnableFeignClients
 public class SecurityMain {
     public static void main(String[] args) {
-        System.out.println("Hello world!");
+        SpringApplication.run(SecurityMain.class, args);
     }
 }