+++ CHANGELOG.md
@@ -11,7 +11,12 @@
 - #83 - Generate test data for PostomatRestControllerIT
 - #12 - Order entity
 - #61 - Front CRUD for Passport
-- #101 - added fuzzy search for product
+- #92 - Security service
+- #101 - Add fuzzy search for product
+- #82 - Generate test data for PickupPointRestController
+- #63 - Front CRUD for PickupPoint
+- #77 - oauth через гугл
+- #66 - Front CRUD for Product
 
 ## [1.1.0] - 8.09.2023
 - #5 - added user entity